Update to meet all mid-sprint goals
Fixed weird button issues, Fixed modal, added updating character count, added updating splitting notification, fixed pictures, fixed a few other minor CSS issues.

The OurModal document's contents was moved to the Body document. This is because Body needs access to some of the state values in the modal. It would be better style to split them up and use a parent to pass states to the two, but that will take a little time. It may be easier to implement auto-splitting first, on the one document, and then splitting it up afterwards.